% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{search_sentinel}
\alias{search_sentinel}
\title{Search Sentinel API
https://planetarycomputer.microsoft.com/dataset/sentinel-2-l2a}
\usage{
search_sentinel(bbox, date_range)
}
\description{
Search Sentinel API
https://planetarycomputer.microsoft.com/dataset/sentinel-2-l2a
}
